// src/enums/EStoreType.ts
export enum ETypeSets {
  All = "Всі",
  GreenTariff = "Зелений тариф",
  AutonomousStations = "Автономні станції",
  BackupPower = "Резервне живлення",
  HybridUPS = "Гібридні безперебійні системи",
  ForEnterprises = "Для підприємств",
}
export enum ETypeComponents {
  All = "Всі",
  Panels = "Панелі",
  Inverters = "Інвертори",
  Batteries = "Акумулятори",
  MountingHardware = "Кріплення",
  Components = "Комплектуючі",
  ChargingStations = "Зарядні станції",
}